[BR-117] : AlbaGPT <-> Alba Manager간 ROS2 토픽 전송 테스트 완료
