+++ alba_gpt/AlbaGPT_main/src/albagpt_client/package.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0"?>
 <?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="3">
-  <name>obstacle_receiver</name>
+  <name>albagpt_client</name>
   <version>0.0.0</version>
   <description>TODO: Package description</description>
   <maintainer email="jungseonglian@sju.ac.kr">addinedu</maintainer>
