+++ alba_gpt/AlbaGPT_main/src/albagpt_client/setup.cfg
@@ -1,0 +1,4 @@
+[develop]
+script_dir=$base/lib/albagpt_client
+[install]
+install_scripts=$base/lib/albagpt_client
