+++ alba_gpt/AlbaGPT_main/src/albagpt_server/albagpt_server/AlbaGPT_function.py
@@ -1,18 +1,12 @@
 from langchain_openai import ChatOpenAI
 from langchain_core.prompts import PromptTemplate
 from langchain.chains import LLMChain
-from langchain_teddynote.models import MultiModal
 from dotenv import load_dotenv
-from AlbaGPT_main import alba_task_type_list, dynamic_object_list
 
+from albagpt_server.config import alba_task_type_list, dynamic_object_list
 import re
-import csv
-import cv2
-import os
-import numpy as np
-import uuid
-import time
 import logging
+import csv
 
 load_dotenv()
 
@@ -43,8 +37,6 @@ def alba_task_discriminator(user_query, memory, alba_task_type_list=alba_task_ty
     for idx in range(len(memory))
     if memory['messages'][idx]['answer'] not in ["채팅 서버에 연결할 수 없습니다. 잠시 후 다시 시도해주세요.", "None"]
     )
-
-    logger.info(chat_history)
 
     task_example_csv_path = './contents/example/task_example.csv' # 유저 프롬프트에 따른 분류된 task를 정리한 csv 파일
     task_example_list = []
@@ -108,6 +100,9 @@ def alba_task_discriminator(user_query, memory, alba_task_type_list=alba_task_ty
         prompt=discriminator_template
     )
 
+    logger.info(llm)
+    logger.info(chain)
+
     discriminated_task = chain.invoke({
         "user_query": user_query,
         "chat_history": chat_history,
@@ -116,6 +111,8 @@ def alba_task_discriminator(user_query, memory, alba_task_type_list=alba_task_ty
     })['text']
 
     discriminated_task = discriminated_task.replace(".", "").strip()
+
+    logger.info(discriminated_task)
     
     return discriminated_task
 
@@ -287,8 +284,6 @@ def generate_alba_maintenance_response(user_query, memory):
         "maintenance_example" : maintenance_example
     })['text']
 
-    print(alba_maintenance_response)
-
     return alba_maintenance_response
 
 def detect_object_obstacles(detected_objects):
@@ -307,91 +302,59 @@ def detect_object_obstacles(detected_objects):
 
     return detected_object_list
 
-def generate_alba_take_picture_response(user_query, memory, shared_dict, robot_id):
+def generate_alba_take_picture_response(user_query, memory, object_info):
     """
     alba_task_discriminator 함수로 구분된 task가 'TAKE_PICTURE'인 경우 user_query에 대해 해당하는 알바 봇으로부터 수신되는 영상의 이미지 프레임을 토대로 상황을 해석해 반환해주는 함수입니다.
 
-    Returns : alba_take_picture_response (str), image_path(str or None)
+    Returns : alba_take_picture_response (str)
     """
-    image_dir = './contents/image'
 
-    if not os.path.exists(image_dir) :
-        os.makedirs(image_dir)
-
-    image_path = os.path.join(image_dir, str(uuid.uuid4().hex) + '_' + time.strftime('%Y-%m-%d %H-%M-%S') + '.jpg')
-
-    shared_id_dict = shared_dict.get(robot_id)
-    
-    if shared_id_dict :
-        shared_id_dict = shared_dict[robot_id]
-        
-        ip_port = shared_id_dict["ip_port"]
-        shared_sub_dict = shared_dict[ip_port]
-        encoded_latest_frame = shared_sub_dict["latest_frame"]
-        detected_objects = shared_sub_dict["detected_object"]
-        
-        detected_object_list = detect_object_obstacles(detected_objects)
-        logger.info(f"💡 Detected Objects : {detected_object_list}")
-            
-        np_data = np.frombuffer(encoded_latest_frame, dtype=np.uint8)
-        decoded_detected_image = cv2.imdecode(np_data, cv2.IMREAD_COLOR)
-
-        if decoded_detected_image is None:
-            raise ValueError("❌ Failed to decode image from decoded_detected_image")
-
-        cv2.imwrite(image_path, decoded_detected_image)
-        logger.info(f"📷 Image successfully saved to {image_path}")
-
-        object_info = ', '.join(detected_object_list)
-
-        if not object_info: # 감지된 동적 장애물이 없다면
-            alba_take_picture_response = "아무 문제 없이 임무를 잘 수행하고 있어요 🤗"
-        else :
-            obstacle_example_path = './contents/example/obstacle_example.txt' # TAKE PICTURES 상황에 대한 응답을 정리한 텍스트 파일
-            with open(obstacle_example_path, 'r', encoding='utf-8') as obstacle_example_file:
-                # 파일 내용을 한 번에 읽고 strip()으로 양쪽 공백 제거
-                obstacle_meetup_example = obstacle_example_file.read().strip()
-
-            chat_history = "\n".join(
-            f"user_query: {memory['messages'][idx]['question']}, response: {memory['messages'][idx]['answer']}"
-            for idx in range(len(memory))
-            if memory['messages'][idx]['answer'] not in ["채팅 서버에 연결할 수 없습니다. 잠시 후 다시 시도해주세요.", "None"]
-            )
-
-            user_prompt = f"""
-            아래는 감지된 사물 목록입니다: {object_info}
-            이 사물들은 경로를 가로막고 있습니다.
-
-            당신의 임무는 다음과 같습니다:
-            1. "{object_info} {obstacle_meetup_example}"와 같은 식으로 대답하여야 합니다. 이때 {object_info}는 한국어로 번역해주어 말을 해야 합니다.
-            2. {object_info}가 '사람'인 경우에도 그냥 "{object_info} {obstacle_meetup_example}"와 같은 식으로 대답하기만 하면 됩니다.
-            3. 중복되는 {object_info}에 대해서는 한 번만 대답합니다.
-            4. 답변은 1-3문장 이내로 간결하게 수행합니다.
-            """
-            multimodal_system_prompt = f"""
-            당신은 레스토랑에서 업무를 하는 '핑키'라는 이름을 가진 모바일 로봇입니다
-            지금까지 당신과 레스토랑 오너가 나눈 대화는 다음과 같습니다 : {chat_history}
-
-            당신은 레스토랑 오너의 {user_query}에 대해 친절하고 상냥하게, 1~3줄 이내로 너무 길지 않게 답변하여야 합니다.   
-
-            [조건]
-            1. response_text는 {user_query}에 대한 string형 답변입니다.
-            2. {obstacle_meetup_example} 중 {chat_history}에 있는 response와는 다른 답변으로 랜덤하게 하나를 선택해 반환해주세요.
-            3. 위의 조건 이외에 다른 정수, 단어, 문장은 생성하지 않습니다.
-            """
-
-            # 객체 생성
-            llm = ChatOpenAI(
-                temperature=0.4,  # 창의성 (0.0 ~ 2.0)
-                max_tokens=2048,  # 최대 토큰수
-                model_name="gpt-4o",  # 모델명
-            )
-
-            multimodal_llm = MultiModal(llm, system_prompt=multimodal_system_prompt, user_prompt=user_prompt)
-            alba_take_picture_response = multimodal_llm.invoke(image_path)
+    if not object_info: # 감지된 동적 장애물이 없다면
+        alba_take_picture_response = "아무 문제 없이 임무를 잘 수행하고 있어요 🤗"
     else :
-        logger.warning(f"🚫 [ERROR] : AlbaBot {robot_id} camera is not on")
-        alba_take_picture_response = f"현재 알바봇 {robot_id}의 카메라가 꺼져 있어요 😢"
-        image_path = None
+        obstacle_example_path = './contents/example/obstacle_example.txt' # TAKE PICTURES 상황에 대한 응답을 정리한 텍스트 파일
+        with open(obstacle_example_path, 'r', encoding='utf-8') as obstacle_example_file:
+            # 파일 내용을 한 번에 읽고 strip()으로 양쪽 공백 제거
+            obstacle_meetup_example = obstacle_example_file.read().strip()
 
-    return alba_take_picture_response, image_path
+        chat_history = "\n".join(
+        f"user_query: {memory['messages'][idx]['question']}, response: {memory['messages'][idx]['answer']}"
+        for idx in range(len(memory))
+        if memory['messages'][idx]['answer'] not in ["채팅 서버에 연결할 수 없습니다. 잠시 후 다시 시도해주세요.", "None"]
+        )
+
+        alba_camera_on_prompt = """
+        당신은 레스토랑에서 업무를 하는 '핑키'라는 이름을 가진 모바일 로봇입니다
+        지금까지 당신과 레스토랑 오너가 나눈 대화는 다음과 같습니다 : {chat_history}
+
+        아래는 감지된 사물 목록입니다: {object_info}
+        이 사물들은 경로를 가로막고 있습니다.
+
+        당신의 임무는 다음과 같습니다:
+        1. "{object_info} {obstacle_meetup_example}"와 같은 식으로 대답하여야 합니다. 이때 {object_info}는 한국어로 번역해주어 말을 해야 합니다.
+        2. response_text는 {user_query}에 대한 string형 답변입니다.
+        3. {object_info}가 '사람'인 경우에도 그냥 "{object_info} {obstacle_meetup_example}"와 같은 식으로 대답하기만 하면 됩니다.
+        4. 위의 조건 이외에 다른 정수, 단어, 문장은 생성하지 않습니다.
+        """
+
+        # 객체 생성
+        llm = ChatOpenAI(
+            temperature=0.4,  # 창의성 (0.0 ~ 2.0)
+            max_tokens=2048,  # 최대 토큰수
+            model_name="gpt-4o",  # 모델명
+        )
+
+        chain = LLMChain(
+        llm=llm,
+        prompt=alba_camera_on_prompt,
+        verbose=True
+        )
+
+        alba_take_picture_response = chain.invoke({
+            "user_query": user_query,
+            "chat_history": chat_history,
+            "object_info" : object_info,
+            "obstacle_meetup_example": obstacle_meetup_example
+        })['text']
+
+    return alba_take_picture_response
