+++ alba_gpt/AlbaGPT_main/src/albagpt_server/albagpt_server/config.py
@@ -1,0 +1,2 @@
+alba_task_type_list = ["GREETINGS", "TAKE_PICTURE", "MAINTENANCE"]
+dynamic_object_list = ["person", "bird", "cat", "dog"]
