+++ alba_gpt/AlbaGPT_main/src/albagpt_server/package.xml
@@ -1,14 +1,14 @@
 <?xml version="1.0"?>
 <?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="3">
-  <name>obstacle_instructor</name>
+  <name>albagpt_server</name>
   <version>0.0.0</version>
   <description>TODO: Package description</description>
   <maintainer email="jungseonglian@sju.ac.kr">addinedu</maintainer>
   <license>TODO: License declaration</license>
 
   <depend>rclpy</depend>
-  <depend>std_msgs</depend>
+  <depend>stdmsgs</depend>
 
   <test_depend>ament_copyright</test_depend>
   <test_depend>ament_flake8</test_depend>
