+++ alba_gpt/AlbaGPT_main/src/albagpt_server/setup.cfg
@@ -1,0 +1,4 @@
+[develop]
+script_dir=$base/lib/albagpt_server
+[install]
+install_scripts=$base/lib/albagpt_server
