--- alba_gpt/AlbaGPT_main/src/obstacle_instructor/setup.cfg
@@ -1,4 +1,0 @@
-[develop]
-script_dir=$base/lib/obstacle_instructor
-[install]
-install_scripts=$base/lib/obstacle_instructor
