--- alba_gpt/AlbaGPT_main/src/obstacle_receiver/setup.cfg
@@ -1,4 +1,0 @@
-[develop]
-script_dir=$base/lib/obstacle_receiver
-[install]
-install_scripts=$base/lib/obstacle_receiver
